fix: move shell detection logic to util
This is now used by both activate and deactivate, so move it off
ActivateOptions into util
