+++ cli/flox/src/utils/mod.rs
@@ -7,6 +7,7 @@ use flox_core::util::default_nix_env_vars;
 
 pub mod active_environments;
 pub mod colors;
+pub mod detect_shell;
 pub mod dialog;
 pub mod didyoumean;
 pub mod errors;
